+++ packages/product-feed-zbozi/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [CONTRIBUTING.md](CONTRIBUTING.md) (@vitek-rostislav)
 
 ## [0.4.0] - 2017-09-12
 ### Added