+++ packages/product-feed-zbozi/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.4.1]
 ### Added
 - [CONTRIBUTING.md](CONTRIBUTING.md) (@vitek-rostislav)
 - [template for github pull requests](docs/PULL_REQUEST_TEMPLATE.md) (@vitek-rostislav)
@@ -55,7 +57,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Extracted Zboží.cz product feed plugin from [Shopsys Framework](http://www.shopsys-framework.com/) (@PetrHeinz)
 - `.travis.yml` file with Travis CI configuration (@PetrHeinz)
 
-[Unreleased]: https://github.com/shopsys/product-feed-zbozi/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/shopsys/product-feed-zbozi/compare/v0.4.1...HEAD
+[0.4.1]: https://github.com/shopsys/product-feed-zbozi/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/shopsys/product-feed-zbozi/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/shopsys/product-feed-zbozi/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/shopsys/product-feed-zbozi/compare/v0.1.0...v0.2.0